backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.34
pydantic==2.9.2
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary==2.9.9
python-dotenv==1.0.0
# Add these if not already present
python-multipart
aiofiles
mediapipe>=0.9.0
opencv-python>=4.5.0

